git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk@372 73e77b4c-caa6-f847-a29a-24ab75ae54b6

+++ Facilities/Prevalence/Castle.Facilities.Prevalence/PrevalenceFacility.cs
@@ -120,7 +120,7 @@ namespace Castle.Facilities.Prevalence
 
 		private void ConfigureSnapshot(IConfiguration engineConfig, IDictionary properties)
 		{
-			int period = GetSnapshotPeriod(engineConfig);
+			float period = GetSnapshotPeriod(engineConfig);
 
 			if (RequiresSnapshots(period))
 			{
@@ -137,11 +137,11 @@ namespace Castle.Facilities.Prevalence
 			}
 		}
 
-		private int GetSnapshotPeriod(IConfiguration engineConfig)
+		private float GetSnapshotPeriod(IConfiguration engineConfig)
 		{
 			try
 			{
-				return Convert.ToInt32(engineConfig.Attributes["snapshotHoursPeriod"]);
+				return Convert.ToSingle(engineConfig.Attributes["snapshotHoursPeriod"]);
 			}
 			catch (FormatException e)
 			{
@@ -149,7 +149,7 @@ namespace Castle.Facilities.Prevalence
 			}
 		}
 
-		private bool RequiresSnapshots(int period)
+		private bool RequiresSnapshots(float period)
 		{
 			return period > 0;
 		}
@@ -184,7 +184,7 @@ namespace Castle.Facilities.Prevalence
 
 		protected void TakeSnapshotIfRequired(IConfiguration engineConfig)
 		{
-			int period = GetSnapshotPeriod(engineConfig);
+			float period = GetSnapshotPeriod(engineConfig);
 
 			if (RequiresSnapshots(period))
 			{
